+++ src/interpreter/executor/calls/int_methods.rs
@@ -56,7 +56,10 @@ impl Interpreter {
             call_env.define(param_name.clone(), Value::Int(i));
 
             match self.execute_block(&func.body, call_env)? {
-                ControlFlow::Return(_) | ControlFlow::Normal(_) | ControlFlow::Continue => {}
+                ControlFlow::Return(_)
+                | ControlFlow::Normal(_)
+                | ControlFlow::Continue
+                | ControlFlow::Break => {}
                 ControlFlow::Throw(_) => {
                     return Err(RuntimeError::new("Exception in int.times", span));
                 }
@@ -100,7 +103,10 @@ impl Interpreter {
             call_env.define(param_name.clone(), Value::Int(i));
 
             match self.execute_block(&func.body, call_env)? {
-                ControlFlow::Return(_) | ControlFlow::Normal(_) | ControlFlow::Continue => {}
+                ControlFlow::Return(_)
+                | ControlFlow::Normal(_)
+                | ControlFlow::Continue
+                | ControlFlow::Break => {}
                 ControlFlow::Throw(_) => {
                     return Err(RuntimeError::new("Exception in int.upto", span));
                 }
@@ -145,7 +151,10 @@ impl Interpreter {
             call_env.define(param_name.clone(), Value::Int(i));
 
             match self.execute_block(&func.body, call_env)? {
-                ControlFlow::Return(_) | ControlFlow::Normal(_) | ControlFlow::Continue => {}
+                ControlFlow::Return(_)
+                | ControlFlow::Normal(_)
+                | ControlFlow::Continue
+                | ControlFlow::Break => {}
                 ControlFlow::Throw(_) => {
                     return Err(RuntimeError::new("Exception in int.downto", span));
                 }